# Closes #1241: Use `computeOnSegments` for `substringSearch`