--- a/src/SegmentedComputation.chpl
+++ b/src/SegmentedComputation.chpl
@@ -44,6 +44,7 @@ module SegmentedComputation {
     StringToNumericReturnValidity,
     StringCompareLiteralEq,
     StringCompareLiteralNeq,
+    StringSearch,
   }
   
   proc computeOnSegments(segments: [?D] int, values: [?vD] ?t, param function: SegFunction, type retType, const strArg: string = "") throws {
@@ -89,6 +90,9 @@ module SegmentedComputation {
               }
               when SegFunction.StringCompareLiteralNeq {
                 agg.copy(res[i], stringCompareLiteralNeq(values, start..#len, strArg));
+              }
+              when SegFunction.StringSearch {
+                agg.copy(res[i], stringSearch(values, start..#len, strArg));
               }
               otherwise {
                 compilerError("Unrecognized segmented function");