This PR (Closes #1241):
- Updates `substringSearch` to use `computeOnSegments`

Edit: The NOTE should be resolved thanks to ronawho's suggestion

<details>
  <summary>NOTE: </summary>

The way I've done this recompiles the regex pattern each iteration instead of (I believe) once per task using `with`
```Chapel
forall ... with (var myRegex = _unsafeCompileRegex(pattern))
```
I'm not sure if this is better or worse/if there's a better way to do this. Feedback on this would be appreciated
</details>