@ronawho 2022-04-06
FYI here are perf results for 16-node-cs-hdr for the str-locality test:

```
# before:
./benchmarks/run_benchmarks.py -nl 16 str-locality
Regex searching good locality Average rate = 40.37 GiB/sec
Regex searching poor locality Average rate =  0.07 GiB/sec

# now:
./benchmarks/run_benchmarks.py -nl 16 str-locality
Regex searching good locality Average rate = 17.96 GiB/sec
Regex searching poor locality Average rate = 11.96 GiB/sec

# now with bumped problem size:
./benchmarks/run_benchmarks.py -nl 16 str-locality --size=$((10**8))
Regex searching good locality Average rate = 25.58 GiB/sec
Regex searching poor locality Average rate = 13.08 GiB/sec
```

(and this is what we expect from computeOnSegments, worse performance with good locality, but much, much better performance with poor locality)